Add vectors (#5)
A vector class adds a layer of abstraction for operations with positions, distances, etc, by using 2D Euclidean vectors.

It provides a nice way to perform element-wise operations with vectors by overriding the addition and subtraction operators. It also has a norm property which will return the norm of the vector, and a method to return a normalized vector with the same direction.