Add maps (#7)
A dungeon is just a subclass of tdl's map with a method to generate itself using rooms, which are rectangles that can check for intersections, and connecting them with tunnels. A tile seen by the player will be marked as explored, so the layout is still rendered when not visible, but not the entities in the invisible tiles.

The generation algorithm is a very basic one. It just spawns a certain number of rooms that don't intersect with each other and connects them to the previous one with tunnels. The player is placed at the center of the first room created. This algorithm will probably be changed in the future.

The render function draws tiles to the buffer console only if the FOV has been recomputed, otherwise we can use the same. Then visible entities are drawn on top of the tiles.

The input method handles horizontal, vertical and diagonal movement, as well as fullscreen toggle and quitting the game.

The main function is mostly there for testing purposes but can be used as a base for the actual main function in the future. It sets some game settings constants, initializes needed objects, and enters the game loop, which continuously listens for player input and reacts accordingly.

A wrapper for numpy arrays using vectors is also added to interact with tdl's Map object in a nicer way.