Add items and item managing (#33)
Add items and item managing

This commit introduces items, objects that can be picked up, used, sold,
etc. by the player (and potentially monsters, in the future ?).

Items may or may not have effects, effects are to be implemented later
on but the concept is that effects shall be generic callables that can
be applied to a number of objects, not necessarily locked to items
(think spells, attacks, traps, ...).

It also adds the backpack which is useful for managing items.